+++ bocadillo/__init__.py
@@ -1,11 +1,13 @@
 from .api import API
-from .exceptions import HTTPError, WebSocketDisconnect
+from .errors import HTTPError
 from .media import Media
 from .middleware import Middleware
 from .recipes import Recipe
 from .staticfiles import static
 from .server_sent_events import server_event
 from .views import view
-from .websockets import WebSocket
+from .websockets import WebSocket, WebSocketDisconnect
+from .response import Response
+from .request import Request
 
-__version__ = "0.8.1"
+__version__ = "0.9.1"