+++ bocadillo/api.py
@@ -1,37 +1,42 @@
 import os
-from functools import wraps
-from typing import Any, Dict, List, Optional, Tuple, Type, Union, Callable
+from functools import partial
+from typing import Any, Dict, List, Optional, Type, Union, Callable, Tuple
 
 from starlette.middleware.cors import CORSMiddleware
 from starlette.middleware.gzip import GZipMiddleware
 from starlette.middleware.httpsredirect import HTTPSRedirectMiddleware
+from starlette.middleware.lifespan import LifespanMiddleware
 from starlette.middleware.trustedhost import TrustedHostMiddleware
 from starlette.middleware.wsgi import WSGIResponder
 from starlette.testclient import TestClient
-from starlette.websockets import WebSocketClose
 from uvicorn.main import get_logger, run
 from uvicorn.reloaders.statreload import StatReload
 
-from .app_types import ASGIApp, ASGIAppInstance, WSGIApp, Scope
-from .compat import call_async
+from .app_types import (
+    ASGIApp,
+    ASGIAppInstance,
+    Scope,
+    Receive,
+    Send,
+    EventHandler,
+    ErrorHandler,
+)
+from .compat import WSGIApp
 from .constants import DEFAULT_CORS_CONFIG
-from .error_handlers import ErrorHandler, error_to_text
-from .events import EventsMixin
-from .exceptions import HTTPError
+from .error_handlers import error_to_text
+from .errors import ServerErrorMiddleware, HTTPErrorMiddleware, HTTPError
 from .media import Media
 from .meta import DocsMeta
-from .middleware import Dispatcher
 from .recipes import RecipeBase
 from .redirection import Redirection
 from .request import Request
 from .response import Response
-from .routing import RoutingMixin, RouteMatch, WebSocketRouteMatch
+from .routing import HTTPRouter, WebSocketRouter
 from .staticfiles import static
 from .templates import TemplatesMixin
-from .websockets import WebSocket
 
 
-class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
+class API(TemplatesMixin, metaclass=DocsMeta):
     """The all-mighty API class.
 
     This class implements the [ASGI](https://asgi.readthedocs.io) protocol.
@@ -64,7 +69,7 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
     enable_cors (bool):
         If `True`, Cross Origin Resource Sharing will be configured according
         to `cors_config`. Defaults to `False`.
-        See also [CORS](../topics/http/middleware.md#cors).
+        See also [CORS](../guides/http/middleware.md#cors).
     cors_config (dict):
         A dictionary of CORS configuration parameters.
         Defaults to `dict(allow_origins=[], allow_methods=["GET"])`.
@@ -72,12 +77,12 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
         If `True`, enable HSTS (HTTP Strict Transport Security) and automatically
         redirect HTTP traffic to HTTPS.
         Defaults to `False`.
-        See also [HSTS](../topics/http/middleware.md#hsts).
+        See also [HSTS](../guides/http/middleware.md#hsts).
     enable_gzip (bool):
         If `True`, enable GZip compression and automatically
         compress responses for clients that support it.
         Defaults to `False`.
-        See also [GZip](../topics/http/middleware.md#gzip).
+        See also [GZip](../guides/http/middleware.md#gzip).
     gzip_min_size (int):
         If specified, compress only responses that
         have more bytes than the specified value.
@@ -86,10 +91,8 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
         Determines how values given to `res.media` are serialized.
         Can be one of the supported media types.
         Defaults to `"application/json"`.
-        See also [Media](../topics/http/media.md).
+        See also [Media](../guides/http/media.md).
     """
-
-    _error_handlers: List[Tuple[Type[Exception], ErrorHandler]]
 
     def __init__(
         self,
@@ -106,40 +109,66 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
     ):
         super().__init__(templates_dir=templates_dir)
 
-        self._error_handlers = []
-        self.add_error_handler(HTTPError, error_to_text)
+        # Debug mode defaults to `False` but it can be set in `.run()`.
+        self._debug = False
 
-        self._extra_apps: Dict[str, Any] = {}
+        # Base ASGI app
+        self.asgi = self.dispatch
 
+        # Mounted apps
+        self.apps: Dict[str, Any] = {}
+
+        # Routers
+        self.http_router = HTTPRouter()
+        self.websocket_router = WebSocketRouter()
+
+        # Test client
         self.client = self.build_client()
 
+        # Static files
         if static_dir is not None:
             if static_root is None:
                 static_root = static_dir
             self.mount(static_root, static(static_dir))
 
+        # Media handlers
         self._media = Media(media_type=media_type)
 
-        self._middleware = []
-        self._asgi_middleware = []
+        # HTTP middleware
+        self.exception_middleware = HTTPErrorMiddleware(
+            self.http_router, debug=self._debug
+        )
+        self.server_error_middleware = ServerErrorMiddleware(
+            self.exception_middleware, handler=error_to_text, debug=self._debug
+        )
+        self.add_error_handler(HTTPError, error_to_text)
 
+        # Lifespan middleware
+        self.lifespan_middleware = LifespanMiddleware(self.dispatch_lifespan)
+
+        # ASGI middleware
         if allowed_hosts is None:
             allowed_hosts = ["*"]
         self.add_asgi_middleware(
             TrustedHostMiddleware, allowed_hosts=allowed_hosts
         )
-
         if enable_cors:
-            if cors_config is None:
-                cors_config = {}
-            cors_config = {**DEFAULT_CORS_CONFIG, **cors_config}
+            cors_config = {**DEFAULT_CORS_CONFIG, **(cors_config or {})}
             self.add_asgi_middleware(CORSMiddleware, **cors_config)
-
         if enable_hsts:
             self.add_asgi_middleware(HTTPSRedirectMiddleware)
-
         if enable_gzip:
             self.add_asgi_middleware(GZipMiddleware, minimum_size=gzip_min_size)
+
+    @property
+    def debug(self) -> bool:
+        return self._debug
+
+    @debug.setter
+    def debug(self, debug: bool):
+        self._debug = debug
+        self.exception_middleware.debug = debug
+        self.server_error_middleware.debug = debug
 
     def build_client(self, **kwargs) -> TestClient:
         return TestClient(self, **kwargs)
@@ -161,10 +190,10 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
         """
         if not prefix.startswith("/"):
             prefix = "/" + prefix
-        self._extra_apps[prefix] = app
+        self.apps[prefix] = app
 
     def recipe(self, recipe: RecipeBase):
-        recipe.apply(self)
+        recipe(self)
 
     @property
     def media_type(self) -> str:
@@ -202,21 +231,15 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
         handler (callable):
             The actual error handler, which is called when an instance of
             `exception_cls` is caught.
-            Should accept a `req`, a `res` and an `exc`.
+            Should accept a request, response and exception parameters.
         """
-        self._error_handlers.insert(0, (exception_cls, handler))
+        self.exception_middleware.add_exception_handler(exception_cls, handler)
 
     def error_handler(self, exception_cls: Type[Exception]):
         """Register a new error handler (decorator syntax).
 
-        # Example
-        ```python
-        >>> import bocadillo
-        >>> api = bocadillo.API()
-        >>> @api.error_handler(KeyError)
-        ... def on_key_error(req, res, exc):
-        ...     pass  # perhaps set res.content and res.status_code
-        ```
+        # See Also
+        - [add_error_handler](#add-error-handler)
         """
 
         def wrapper(handler):
@@ -225,26 +248,75 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
 
         return wrapper
 
-    def _find_handler(
-        self, exception_cls: Type[Exception]
-    ) -> Optional[ErrorHandler]:
-        for cls, handler in self._error_handlers:
-            if issubclass(exception_cls, cls):
-                return handler
-        return None
+    def route(self, pattern: str, *, name: str = None, namespace: str = None):
+        """Register a new route by decorating a view.
 
-    async def _handle_exception(
-        self, req: Request, res: Response, exception: Exception
-    ) -> None:
-        # Handle an exception raised during dispatch.
-        # If no handler was registered for the exception, it is raised.
-        handler = self._find_handler(exception.__class__)
-        if handler is None:
-            return await self._handle_exception(req, res, HTTPError(500))
+        # Parameters
+        pattern (str): an URL pattern.
+        methods (list of str):
+            An optional list of HTTP methods.
+            Defaults to `["get", "head"]`.
+            Ignored for class-based views.
+        name (str):
+            An optional name for the route.
+            If a route already exists for this name, it is replaced.
+            Defaults to a snake-cased version of the view's name.
+        namespace (str):
+            An optional namespace for the route. If given, it is prefixed to
+            the name and separated by a colon.
 
-        await call_async(handler, req, res, exception)
-        if res.status_code is None:
-            res.status_code = 500
+        # See Also
+        - [check_route](#check-route) for the route validation algorithm.
+        """
+        return self.http_router.route(
+            pattern=pattern, name=name, namespace=namespace
+        )
+
+    def websocket_route(
+        self,
+        pattern: str,
+        *,
+        value_type: Optional[str] = None,
+        receive_type: Optional[str] = None,
+        send_type: Optional[str] = None,
+        caught_close_codes: Optional[Tuple[int]] = None,
+    ):
+        """Register a WebSocket route by decorating a view.
+
+        # Parameters
+        pattern (str): an URL pattern.
+
+        # See Also
+        - [WebSocket](./websockets.md#websocket) for a description of keyword
+        arguments.
+        """
+        # NOTE: use named keyword arguments instead of `**kwargs` to improve
+        # their accessibility (e.g. for IDE discovery).
+        return self.websocket_router.route(
+            pattern,
+            value_type=value_type,
+            receive_type=receive_type,
+            send_type=send_type,
+            caught_close_codes=caught_close_codes,
+        )
+
+    def url_for(self, name: str, **kwargs) -> str:
+        """Build the URL path for a named route.
+
+        # Parameters
+        name (str): the name of the route.
+        kwargs (dict): route parameters.
+
+        # Returns
+        url (str): the URL path for a route.
+
+        # Raises
+        HTTPError(404) : if no route exists for the given `name`.
+        """
+        route = self.http_router.routes.get(name)
+        if route is None:
+            raise HTTPError(404)
+        return route.url(**kwargs)
 
     def redirect(
         self,
@@ -254,11 +326,12 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
         permanent: bool = False,
         **kwargs,
     ):
-        """Redirect to another route.
+        """Redirect to another HTTP route.
 
         # Parameters
         name (str): name of the route to redirect to.
-        url (str): URL of the route to redirect to, required if `name` is omitted.
+        url (str):
+            URL of the route to redirect to (required if `name` is omitted).
         permanent (bool):
             If `False` (the default), returns a temporary redirection (302).
             If `True`, returns a permanent redirection (301).
@@ -266,10 +339,11 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
             Route parameters.
 
         # Raises
-        Redirection: an exception that will be caught by #API.dispatch().
+        Redirection:
+            an exception that will be caught to trigger a redirection.
 
         # See Also
-        - [Redirecting](../topics/http/redirecting.md)
+        - [Redirecting](../guides/http/redirecting.md)
         """
         if name is not None:
             url = self.url_for(name=name, **kwargs)
@@ -286,9 +360,11 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
             A subclass of `bocadillo.Middleware`.
 
         # See Also
-        - [Middleware](../topics/http/middleware.md)
+        - [Middleware](../guides/http/middleware.md)
         """
-        self._middleware.insert(0, (middleware_cls, kwargs))
+        self.exception_middleware.app = middleware_cls(
+            self.exception_middleware.app, **kwargs
+        )
 
     def add_asgi_middleware(self, middleware_cls, *args, **kwargs):
         """Register an ASGI middleware class.
@@ -298,169 +374,92 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
             A class that complies with the ASGI specification.
 
         # See Also
-        - [Middleware](../topics/http/middleware.md)
+        - [ASGI middleware](../guides/agnostic/asgi-middleware.md)
         - [ASGI](https://asgi.readthedocs.io)
         """
-        self._asgi_middleware.insert(0, (middleware_cls, args, kwargs))
+        self.asgi = middleware_cls(self.asgi, *args, **kwargs)
 
-    def apply_asgi_middleware(self, app: ASGIApp) -> ASGIApp:
-        """Wrap the registered ASGI middleware around an ASGI app.
-
-        # Parameters
-        app (ASGIApp): a callable complying with the ASGI specification.
-
-        # Returns
-        app_with_asgi_middleware (ASGIApp):
-            The result `app = asgi(app, *args, **kwargs)` for
-            each ASGI middleware class.
-
-        # See Also
-        - [add_asgi_middleware](#add-asgi-middleware)
-        """
-        for middleware_cls, args, kwargs in self._asgi_middleware:
-            app: ASGIApp = middleware_cls(app, *args, **kwargs)
-        return app
-
-    async def dispatch(self, req: Request) -> Response:
-        """Dispatch a request and return a response.
+    def on(self, event: str, handler: Optional[EventHandler] = None):
+        """Register an event handler.
 
         # Parameters
-        req (Request): an inbound HTTP request.
+        event (str):
+            Either `"startup"` (when the server boots) or `"shutdown"`
+            (when the server stops).
+        handler (callback, optional):
+            The event handler. If not given, this should be used as a
+            decorator.
 
-        # Returns
-        response (Response): an HTTP response.
+        # Example
 
-        # See Also
-        - [How are requests processed?](../topics/http/routes-url-design.md#how-are-requests-processed) for the dispatch algorithm.
+        ```python
+        @api.on("startup")
+        async def startup():
+            pass
+
+        async def shutdown():
+            pass
+
+        api.on("shutdown", shutdown)
+        ```
         """
-        res = Response(req, media=self._media)
+        if handler is None:
 
-        try:
-            match: RouteMatch = self._router.match(req.url.path)
-            if match is None:
-                raise HTTPError(status=404)
+            def register(func):
+                self.lifespan_middleware.add_event_handler(event, func)
+                return func
 
-            try:
-                await match.route(req, res, **match.params)
-            except Redirection as redirection:
-                res = redirection.response
-
-        except Exception as e:
-            await self._handle_exception(req, res, e)
-
-        return res
-
-    def _convert_exception_to_response(
-        self, dispatch: Dispatcher
-    ) -> Dispatcher:
-        # Wrap call to `dispatch()` to always return an HTTP response.
-
-        @wraps(dispatch)
-        async def inner(req: Request) -> Response:
-            try:
-                res = await dispatch(req)
-            except Exception as exc:
-                res = Response(req, media=self._media)
-                await self._handle_exception(req, res, exc)
-            return res
-
-        return inner
-
-    async def get_response(self, req: Request) -> Response:
-        """Return a response for an incoming request.
-
-        # Parameters
-        req (Request): a Request object.
-
-        # Returns
-        res (Response):
-            a Response object, obtained by going down the middleware chain,
-            calling `dispatch()` and going up the middleware chain.
-
-        # See Also
-        - [dispatch](#dispatch)
-        - [Middleware](../topics/http/middleware.md)
-        """
-        convert = self._convert_exception_to_response
-        dispatch = convert(self.dispatch)
-        for cls, kwargs in self._middleware:
-            middleware = cls(dispatch, **kwargs)
-            dispatch = convert(middleware)
-        return await dispatch(req)
-
-    async def dispatch_websocket(self, scope: Scope, receive, send):
-        match: WebSocketRouteMatch = self._router.match(
-            scope["path"], websocket=True
-        )
-        if match is None:
-            # Close with a 403 error code, as specified in the ASGI spec:
-            # https://asgi.readthedocs.io/en/latest/specs/www.html#close
-            await WebSocketClose(code=403)(receive, send)
+            return register
         else:
-            ws = WebSocket(scope, receive, send, **match.route.ws_kwargs)
-            try:
-                await match.route(ws, **match.params)
-            except BaseException:
-                await ws.ensure_closed(1011)
-                raise
+            self.lifespan_middleware.add_event_handler(event, handler)
+            return handler
 
-    def create_app(self, scope: Scope) -> ASGIAppInstance:
-        """Build and return an instance of the `API`'s own ASGI application.
-
-        # Parameters
-        scope (dict): an ASGI connection scope.
-
-        # Returns
-        asgi (ASGIAppInstance):
-            creates a `Request` and awaits the result of `get_response()`.
-        """
+    def dispatch_lifespan(self, scope: Scope):
+        # Strict implementation of the ASGI lifespan spec.
+        # This is required because the Starlette `LifespanMiddleware`
+        # does not send the `complete` responses.
 
         async def asgi(receive, send):
-            nonlocal scope
-            if scope["type"] == "websocket":
-                await self.dispatch_websocket(scope, receive, send)
-            else:
-                assert scope["type"] == "http"
-                req = Request(scope, receive)
-                res = await self.get_response(req)
-                await res(receive, send)
+            message = await receive()
+            assert message["type"] == "lifespan.startup"
+            await send({"type": "lifespan.startup.complete"})
+
+            message = await receive()
+            assert message["type"] == "lifespan.shutdown"
+            await send({"type": "lifespan.shutdown.complete"})
 
         return asgi
 
-    def find_app(self, scope: Scope) -> ASGIAppInstance:
-        """Return the ASGI application suited to the given ASGI scope.
+    async def dispatch_http(self, receive: Receive, send: Send, scope: Scope):
+        assert scope["type"] == "http"
+        req = Request(scope, receive)
+        res = Response(req, media=self._media)
+        res = await self.server_error_middleware(req, res)
+        await res(receive, send)
+        # Re-raise the exception to allow the server to log the error
+        # and for the test client to optionally re-raise it too.
+        self.server_error_middleware.raise_if_exception()
 
-        The application is chosen according to the following algorithm:
+    async def dispatch_websocket(
+        self, receive: Receive, send: Send, scope: Scope
+    ):
+        await self.websocket_router(scope, receive, send)
 
-        - If `scope` has a `lifespan` type, the lifespan handler is returned.
-        This occurs on server startup and shutdown.
-        - If the scope's `path` begins with any of the prefixes of a mounted
-        sub-app, said sub-app is returned (converting from WSGI to ASGI if
-        necessary).
-        - Otherwise, the `API`'s own ASGI application is returned.
+    def dispatch(self, scope: Scope) -> ASGIAppInstance:
+        if scope["type"] == "websocket":
+            return partial(self.dispatch_websocket, scope=scope)
+        else:
+            assert scope["type"] == "http"
+            return partial(self.dispatch_http, scope=scope)
 
-        # Parameters
-        scope (dict):
-            An ASGI scope.
-
-        # Returns
-        app:
-            An ASGI application instance.
-
-        # See Also
-        - [Lifespan Protocol](https://asgi.readthedocs.io/en/latest/specs/lifespan.html)
-        - [ASGI connection scope](https://asgi.readthedocs.io/en/latest/specs/main.html#connection-scope)
-        - [Events](../topics/agnostic/events.md)
-        - [mount](#mount)
-        - [create_app](#create-app)
-        """
+    def __call__(self, scope: Scope) -> ASGIAppInstance:
         if scope["type"] == "lifespan":
-            return self.handle_lifespan(scope)
+            return self.lifespan_middleware(scope)
 
         path: str = scope["path"]
 
         # Return a sub-mounted extra app, if found
-        for prefix, app in self._extra_apps.items():
+        for prefix, app in self.apps.items():
             if not path.startswith(prefix):
                 continue
             # Remove prefix from path so that the request is made according
@@ -471,7 +470,7 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
             except TypeError:
                 return WSGIResponder(app, scope)
 
-        return self.apply_asgi_middleware(self.create_app)(scope)
+        return self.asgi(scope)
 
     def run(
         self,
@@ -504,8 +503,8 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
             Extra keyword arguments that will be passed to the Uvicorn runner.
 
         # See Also
-        - [Configuring host and port](../topics/api.md#configuring-host-and-port)
-        - [Debug mode](../topics/api.md#debug-mode)
+        - [Configuring host and port](../guides/api.md#configuring-host-and-port)
+        - [Debug mode](../guides/api.md#debug-mode)
         - [Uvicorn settings](https://www.uvicorn.org/settings/) for all
         available keyword arguments.
         """
@@ -524,6 +523,7 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
             port = 8000
 
         if debug:
+            self.debug = True
             reloader = StatReload(get_logger(log_level))
             reloader.run(
                 run,
@@ -532,12 +532,9 @@ class API(TemplatesMixin, RoutingMixin, EventsMixin, metaclass=DocsMeta):
                     "host": host,
                     "port": port,
                     "log_level": log_level,
-                    "debug": debug,
+                    "debug": self.debug,
                     **kwargs,
                 },
             )
         else:
             _run(self, host=host, port=port, **kwargs)
-
-    def __call__(self, scope: Scope) -> ASGIAppInstance:
-        return self.find_app(scope)