+++ docs/.vuepress/enhanceApp.js
@@ -1,0 +1,6 @@
+export default ({ router }) => {
+    router.addRoutes([
+        // After renaming "Topics" section to "Guides"
+        { path: '/topics/', redirect: '/guides/' },
+    ])
+}