+++ tests/test_http_middleware.py
@@ -7,19 +7,19 @@ from bocadillo import API, Middleware, HTTPError
 
 
 @contextmanager
-def build_middleware(expect_kwargs=None, sync=False):
+def build_middleware(expect_kwargs=None, sync=False, expect_call_after=True):
     called = {"before": False, "after": False}
     kwargs = None
 
     class SetCalled(Middleware):
-        def __init__(self, dispatch, **kw):
+        def __init__(self, app, **kw):
             nonlocal kwargs
             kwargs = kw
-            super().__init__(dispatch, **kw)
+            super().__init__(app, **kw)
 
         if sync:
 
-            def before_dispatch(self, req):
+            def before_dispatch(self, req, res):
                 nonlocal called
                 called["before"] = True
 
@@ -29,7 +29,7 @@ def build_middleware(expect_kwargs=None, sync=False):
 
         else:
 
-            async def before_dispatch(self, req):
+            async def before_dispatch(self, req, res):
                 nonlocal called
                 await sleep(0.01)
                 called["before"] = True
@@ -42,7 +42,7 @@ def build_middleware(expect_kwargs=None, sync=False):
     yield SetCalled
 
     assert called["before"] is True
-    assert called["after"] is True
+    assert called["after"] is expect_call_after
     if expect_kwargs is not None:
         assert kwargs == expect_kwargs
 
@@ -70,8 +70,8 @@ def test_can_pass_extra_kwargs(api: API):
         api.client.get("/")
 
 
-def test_callbacks_are_called_if_method_not_allowed(api: API):
-    with build_middleware() as middleware:
+def test_only_before_dispatch_is_called_if_method_not_allowed(api: API):
+    with build_middleware(expect_call_after=False) as middleware:
         api.add_middleware(middleware)
 
         @api.route("/")
@@ -95,16 +95,16 @@ def test_callbacks_can_be_sync(api: API):
 
 
 @pytest.mark.parametrize("when", ["before", "after"])
-def test_errors_raised_in_callback_return_500_error(api: API, when):
+def test_errors_raised_in_callback_are_handled(api: API, when):
     class CustomError(Exception):
         pass
 
     @api.error_handler(CustomError)
     def handle_error(req, res, exception):
-        pass  # mute exception
+        res.text = "gotcha!"
 
     class MiddlewareWithErrors(Middleware):
-        async def before_dispatch(self, req):
+        async def before_dispatch(self, req, res):
             if when == "before":
                 raise CustomError
 
@@ -118,8 +118,10 @@ def test_errors_raised_in_callback_return_500_error(api: API, when):
     async def index(req, res):
         pass
 
-    response = api.client.get("/")
-    assert response.status_code == 500
+    client = api.build_client(raise_server_exceptions=False)
+    r = client.get("/")
+    assert r.status_code == 200
+    assert r.text == "gotcha!"
 
 
 def test_middleware_uses_registered_http_error_handler(api: API):
@@ -129,7 +131,7 @@ def test_middleware_uses_registered_http_error_handler(api: API):
         res.text = "Foo"
 
     class NopeMiddleware(Middleware):
-        async def before_dispatch(self, req):
+        async def before_dispatch(self, req, res):
             raise HTTPError(401)
 
     api.add_middleware(NopeMiddleware)
@@ -141,3 +143,21 @@ def test_middleware_uses_registered_http_error_handler(api: API):
     response = api.client.get("/")
     assert response.status_code == 401
     assert response.text == "Foo"
+
+
+def test_return_response_in_before_hook(api: API):
+    class NopeMiddleware(Middleware):
+        async def before_dispatch(self, req, res):
+            res.text = "Foo"
+            return res
+
+    api.add_middleware(NopeMiddleware)
+
+    @api.route("/")
+    async def index(req, res):
+        # Should not be called
+        assert False
+
+    r = api.client.get("/")
+    assert r.status_code == 200
+    assert r.text == "Foo"